* #3: Moved the files and added documentation
