+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 application-properties.yml
+build
