--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import exasol_rest_api "main/cmd/exasol-rest-api"
-
-func main() {
-	exasol_rest_api.Run()
-}
