+++ pkg/exasol-rest-api/application.go
