+++ pkg/exasol-rest-api/application_properties.go
