+++ pkg/exasol-rest-api/logger.go
