+++ pkg/exasol-rest-api/properties_file_parser.go
