+++ pkg/exasol-rest-api/query.go
