+++ pkg/exasol-rest-api/request_types.go
