+++ pkg/exasol-rest-api/websocket_connection.go
